refactoring constructors has become rewriting the entire core...
