+++ pleiades/patches.py
@@ -1,0 +1,9 @@
+import matplotlib as mpl
+from abc import ABCMeta
+
+
+class PatchMixin(metaclass=ABCMeta):
+    """A mixin class for adding patch plotting functionality to
+    CurrentFilamentSet objects
+    """
+    pass
